Refactor: Member Client 분리(#32)
인터페이스로 분리
